+++ public/index.html
@@ -6,11 +6,7 @@
   <title>Goose</title>
   <link rel="icon" href="https://i.imgur.com/rl1iuln.png"></link>
   <meta name="viewport" content="width=device-width, initial-scale=1">
-<<<<<<< HEAD
 <script defer src="/js/dist/vendors-node_modules_css-loader_dist_runtime_api_js-node_modules_css-loader_dist_runtime_sour-c8c4ef.d62a94d19270f58858b6.js?bb8753fe21071f10ebc4"></script><script defer src="/js/dist/App.ec955b113be92640dd89.js?bb8753fe21071f10ebc4"></script><script defer src="/js/dist/main.59bae51ef0effcb507ff.js?bb8753fe21071f10ebc4"></script></head>
-=======
-<script defer src="/js/dist/vendors-node_modules_css-loader_dist_runtime_api_js-node_modules_css-loader_dist_runtime_sour-c8c4ef.d62a94d19270f58858b6.js?bbf9c2fbed730cf4eccc"></script><script defer src="/js/dist/App.a7d3da73d4b545c9f906.js?bbf9c2fbed730cf4eccc"></script><script defer src="/js/dist/main.59bae51ef0effcb507ff.js?bbf9c2fbed730cf4eccc"></script></head>
->>>>>>> dev
 
 <body>
   <div id="app"></div>
