+++ src/pages/HomePage/HomePage.js
@@ -2,38 +2,8 @@ import React, { useState, useEffect } from 'react';
 import styles from './HomePage.module.scss';
 import PostList from '../../components/PostList/PostList';
 import NewPostForm from '../../components/NewPostForm/NewPostForm';
-import { getAllPosts } from '../../utilities/posts-service'
-import { indexUsers, getUser } from '../../utilities/users-service'
 
-export default function HomePage({  user, setUsers }) {  
-    const [posts, setPosts] = useState([]);  
-
-    const fetchPosts = async () => {
-        console.log('fetch posts use effect...')
-        try {
-            const postsData = await getAllPosts();
-            setPosts(postsData);
-        } catch (error) {
-            console.error('There was an error!', error);
-        }
-    }
-    const fetchUsers = async () => {
-        try {
-            const foundUsers = await indexUsers();
-            setUsers(foundUsers);
-        } catch (error) {
-            console.error('Error finding users', error);
-        }
-    }
-    useEffect(() => {
-        fetchPosts();
-    }, []);
-    useEffect(() => {
-        fetchUsers();
-    }, []);
-    
-    
-
+export default function HomePage({ posts, fetchPosts, user }) {                
     return (
         <div className={styles.HomePage}>            
             {
